Move renderGoMobile to go-mobile-card component
