+++ client/my-sites/customer-home/style.scss
@@ -190,15 +190,7 @@
 			}
 		}
 	}
-	&__card-mobile {
-		display: flex;
-		flex-direction: row;
-		align-items: center;
 
-		.get-apps__app-badge {
-			width: 50%;
-		}
-	}
 	&__card-subheader {
 		color: var( --color-text-subtle );
 		font-size: 0.85rem;
@@ -207,26 +199,7 @@
 	&__grow-earn {
 		padding-bottom: 12px;
 	}
-	&__go-mobile.is-single-store {
-		display: flex;
-		padding-bottom: 1rem;
 
-		.customer-home__card-subheader,
-		.get-apps__app-badge img {
-			margin: 0;
-		}
-
-		.get-apps__app-badge {
-			width: auto;
-		}
-
-		.customer-home__card-mobile {
-			margin-left: auto;
-			align-self: center;
-			flex-shrink: 0;
-			padding-left: 1rem;
-		}
-	}
 	&__layout {
 		@include grid-row( 3, 1 );
 		@include breakpoint( '>1040px' ) {
